common header for template helpers
